Allow network node feature switching
